Added an option for ignoring tracking account outflows in the expense calculations.
Added an option to allow four different ways to calculate FI on the budget screen. These are:
1) Static - End of Month
2) Static - Start of Month
3) Dynamic - End of Month
4) Dynamic - Start of Month

The static methods do not update the number based on the displayed budget month.
The dynamic methods DO update the number based on the displayed budget month.

End of Month means the end of the current month or currently displayed budeget month. For the static method, this is effectively to-date.
Start of Month means the last day of last month or the last day of the month prior to the displayed month.